+++ alea/utils.py
@@ -13,6 +13,8 @@ from base64 import b32encode
 from collections.abc import Mapping
 from typing import Any, List, Dict, Tuple, Optional, Union, cast, get_args, get_origin
 
+import h5py
+
 # These imports are needed to evaluate strings
 import numpy  # noqa: F401
 import numpy as np  # noqa: F401
@@ -374,6 +376,44 @@ def add_i_batch(filename: str) -> str:
     return fpat_split[0] + "_{i_batch:d}" + fpat_split[1]
 
 
+def search_filename_pattern(filename: str) -> str:
+    """Return pattern for a given existing filename. This is needed because sometimes the filename
+    is not appended by "_{i_batch:d}". We need to distinguish between the two cases and return the
+    correct pattern.
+
+    Returns:
+        str: existing pattern for filename, either filename or filename w/ inserted "_*"
+
+    """
+    # try to add a * to the filename to read all the files
+    fpat_split = os.path.splitext(filename)
+    _filename = fpat_split[0] + "_*" + fpat_split[1]
+    if len(sorted(glob(_filename))) != 0:
+        pattern = _filename
+    else:
+        pattern = filename
+    filename_list = sorted(glob(pattern))
+    if len(filename_list) == 0:
+        raise ValueError(f"Can not find any output file {filename}!")
+    return pattern
+
+
+def get_metadata(output_filename_pattern: str) -> list:
+    """Get metadata from output files."""
+    output_filename_list = sorted(glob(output_filename_pattern))
+    metadata_list = []
+    for _output_filename in output_filename_list:
+        with h5py.File(_output_filename, "r", libver="latest", swmr=True) as ipt:
+            metadata = dict(
+                zip(
+                    ipt.attrs.keys(),
+                    [json.loads(ipt.attrs[key]) for key in ipt.attrs.keys()],
+                )
+            )
+        metadata_list.append(metadata)
+    return metadata_list
+
+
 def can_expand_grid(variations: dict) -> bool:
     """Check if variations can be expanded into a grid.
 
@@ -449,7 +489,7 @@ def convert_variations(variations: dict, iteration) -> list:
         if not isinstance(v, list):
             raise ValueError(f"variations {k} must be a list, not {v} with {type(v)}")
         variations[k] = expand_grid_dict(v)
-    result = [dict(zip(variations, t)) for t in iteration(*variations.values())]
+    result = [dict(zip(variations, deepcopy(t))) for t in iteration(*variations.values())]
     if result:
         return result
     else: